Hibernate: move enhancers around so we can know when it's done registering enhancers
